NodeEditorApp.tsxをリファクタ
