+++ apps/mooreseditor/src/nodeEditor/hooks/useGraphChangeHandlers.ts
@@ -1,0 +1,86 @@
+import { useCallback } from "react";
+
+import { applyNodeChanges, applyEdgeChanges } from "@xyflow/react";
+
+import { useNodeEditorContext } from "../context/NodeEditorContext";
+import { removeRecipesFromJsonData } from "../utils/recipeCleanup";
+
+import type { Column } from "../../hooks/useJson";
+import type { SchemaMeta } from "../utils/schemaMeta";
+import type {
+  OnNodesChange,
+  OnEdgesChange,
+  Node as ReactFlowNode,
+  Edge as ReactFlowEdge,
+} from "@xyflow/react";
+
+interface UseGraphChangeHandlersParams {
+  schemaMetas: Map<string, SchemaMeta>;
+  setJsonData: React.Dispatch<React.SetStateAction<Column[]>>;
+  onMarkDirty: () => void;
+}
+
+export function useGraphChangeHandlers({
+  schemaMetas,
+  setJsonData,
+  onMarkDirty,
+}: UseGraphChangeHandlersParams) {
+  const { state, dispatch } = useNodeEditorContext();
+
+  const onNodesChange: OnNodesChange = useCallback(
+    (changes) => {
+      const newNodes = applyNodeChanges(changes, state.nodes);
+      dispatch({ type: "SET_NODES", nodes: newNodes });
+      onMarkDirty();
+    },
+    [state.nodes, dispatch, onMarkDirty],
+  );
+
+  const onEdgesChange: OnEdgesChange = useCallback(
+    (changes) => {
+      const removeChanges = changes.filter((c) => c.type === "remove");
+      if (removeChanges.length > 0) {
+        const edgesToRemove = removeChanges
+          .map((c) => state.edges.find((e) => e.id === c.id))
+          .filter((e): e is ReactFlowEdge => e != null);
+        if (edgesToRemove.length > 0) {
+          setJsonData((prev) =>
+            removeRecipesFromJsonData(edgesToRemove, state.edges, prev, schemaMetas),
+          );
+        }
+      }
+
+      const newEdges = applyEdgeChanges(changes, state.edges);
+      dispatch({ type: "SET_EDGES", edges: newEdges });
+      onMarkDirty();
+    },
+    [state.edges, dispatch, setJsonData, onMarkDirty, schemaMetas],
+  );
+
+  const handleNodeSelect = useCallback(
+    (node: ReactFlowNode | null) => {
+      dispatch({ type: "SET_SELECTED_NODE", nodeId: node?.id ?? null });
+    },
+    [dispatch],
+  );
+
+  const handleViewportChange = useCallback(
+    (viewport: { x: number; y: number; zoom: number }) => {
+      dispatch({ type: "SET_VIEWPORT", viewport });
+    },
+    [dispatch],
+  );
+
+  const handleMarkDirty = useCallback(() => {
+    onMarkDirty();
+    dispatch({ type: "SET_DIRTY", dirty: true });
+  }, [onMarkDirty, dispatch]);
+
+  return {
+    onNodesChange,
+    onEdgesChange,
+    handleNodeSelect,
+    handleViewportChange,
+    handleMarkDirty,
+  };
+}
